can find closest metro stations

+++ db/migrate/20150309202339_create_metro_stations.rb
@@ -3,8 +3,8 @@ class CreateMetroStations < ActiveRecord::Migration
     create_table :metro_stations do |t|
       t.string    :station_code
       t.string    :station_name
-      t.integer   :longitude
-      t.integer   :latitude
+      t.decimal   :longitude
+      t.decimal   :latitude
       
       t.timestamps null: false
     end
